refactor: move faqs.ts into lib/data directory
